Modified users list component according to remarks from team members. Renamed some files according to JS convention.
