+++ MedicalDocManagment.WebUI/Content/Site.css
@@ -17,13 +17,6 @@ textarea {
 }
 
 
-
-h1 {
-  color: #369;
-  font-family: Arial, Helvetica, sans-serif;
-  font-size: 250%;
-}
-
 h3 {
     text-align: center;
 }
@@ -33,12 +26,7 @@ table {
     border: 1px solid black;
 }
 
-th {
-    text-align: center;
-    border: 1px solid black;
-}
-
-td {
+thead, th, tbody, td {
     text-align: center;
     border: 1px solid black;
 }
