--- MedicalDocManagment.WebUI/app/users.list.ts
@@ -1,27 +1,0 @@
-﻿import { Component } from '@angular/core';
-import { Response } from '@angular/http';
-
-
-@Component({
-    selector: 'emptyContent',
-    templateUrl: './app/views/users.list.html'
-})
-
-export class UsersListComponent {
-    users = this.getUsersList();
-
-    ngOnInit() {
-        this.getUsersList()
-    }
-
-    getUsersList() {
-        // mock-up
-       return [
-            { username: "Іван", email: "ivan@lala.net", position: "лікар", role: "якась роль"},
-            { username: "Сергій", email: "Sergiy@lala.net", position: "лікар", role: "якась роль"},
-            { username: "Степан", email: "Stepan@lala.net", position: "лікар", role: "якась роль"}
-        ]
-    }
-
-}
-
