+++ MedicalDocManagment.WebUI/app/views/user-userslist-component.html
@@ -1,0 +1,5 @@
+﻿<td>{{user.username}}</td>
+<td>{{user.avatar}}</td>
+<td>{{user.email}}</td>
+<td>{{user.position}}</td>
+<td>{{user.role.role}}</td>
